Refactor authentication
- Clear refresh token cookie when logging out from all devices
- Added cookie clearance to logoutAllPlace function
- Enhanced logoutAllPlace to remove refreshToken cookie
- Ensure refreshToken cookie is cleared on logout from all places

Note: login, autologin, logout and logoutAllPlace (CreateNewBook)is completed
