+++ backend/controllers/tokensControllers.js
@@ -1,4 +1,5 @@
 import Token from "../models/Token.js";
+import { verifyRefreshToken } from "../middleware/jwt.js";
 
 const getAllRefreshToken = async (req, res, next) => {
   try {
@@ -11,28 +12,46 @@ const getAllRefreshToken = async (req, res, next) => {
   }
 };
 
-const deleteTokenById = async (req, res, next) => {
+const logout = async (req, res, next) => {
+  const refreshToken = req.cookies.refreshToken;
+  if (!refreshToken) {
+    return res.status(400).json({ message: "No refresh token provided" });
+  }
   try {
-    const id = req.params.id;
-    await Token.deleteById(id);
-    res.status(200).json({ message: "Token is deleted successfully." });
+    const decodedToken = verifyRefreshToken(refreshToken);
+    const userId = decodedToken.user.id;
+    const tokenId = await Token.findIdByToken(refreshToken);
+    const tokenInDb = await Token.findRefreshTokenForUser(userId, refreshToken);
+    if (!tokenInDb) {
+      return res.status(401).json({ message: "Invalid refresh token" });
+    }
+    await Token.deleteById(tokenId);
+    res.clearCookie("refreshToken");
+    res.status(200).json({ message: "Successfully logged out" });
   } catch (error) {
-    console.log(error);
+    console.error(error);
     res.status(500).json({ message: "Server Error", error: error.message });
     next(error);
   }
 };
 
-const deleteAllTokenByUserId = async (req, res, next) => {
+const logoutAllPlace = async (req, res, next) => {
   try {
     const userId = req.params.id;
 
     await Token.deleteByUserId(userId);
-    res.status(200).json({ message: "All Token is deleted successfully." });
+    res.clearCookie("refreshToken");
+    res
+      .status(200)
+      .json({
+        message:
+          "All tokens are deleted successfully and current device cookie cleared.",
+      });
   } catch (error) {
     console.log(error);
     res.status(500).json({ message: "Server Error", error: error.message });
     next(error);
   }
 };
-export default { getAllRefreshToken, deleteTokenById, deleteAllTokenByUserId };
+
+export default { getAllRefreshToken, logout, logoutAllPlace };
